Refactor CLI workspace commands into submodules
